+++ src/frontendGUI/gameButtons/updateGameHandle.java
@@ -63,7 +63,7 @@ public class updateGameHandle {
 			} else if ((oldInfo[i] != null || newInfo[i] != null) && (i == 9 || i == 10 || i == 11 || i == 12)) {
 				switch (i) {
 					case 9:
-						String[] jrb1 = {jrb[0]!=null?jrb[0]:"Not played", jrb[1]!=null?jrb[1]:"In progress", jrb[2]!=null?jrb[2]:"Finish", jrb[3]!=null?jrb[3]:"100% Finished"};
+						String[] jrb1 = {jrb[0]!=null?jrb[0]:"Not played", jrb[1]!=null?jrb[1]:"In progress", jrb[2]!=null?jrb[2]:"Finish", jrb[3]!=null?jrb[3]:"100% Finished", "Dropped"};
 						radioButtons("progress", jrb1, defaultValues.infoProgress, true, oldInfo[i]);
 						panel.add(new JLabel(jla[i]!=null?jla[i]:"N/A")); panel.add(howFarUserPlayedPanel);
 						break;
@@ -144,7 +144,7 @@ public class updateGameHandle {
 		JPanel buttonPanel = new JPanel();
 
 		if (isRB) { 
-			for (int i = 0; i < jrbArray.length; i++) {
+			for (int i = 0; i < action.length; i++) {
 				JRadioButton button = new JRadioButton(jrbArray[i], i == 0); 
 				button.setActionCommand(action[i]);
 				allButtons.add(button);
@@ -153,7 +153,7 @@ public class updateGameHandle {
 			}
 		}
 		else { 
-			for (int i = 0; i < jrbArray.length; i++) {
+			for (int i = 0; i < action.length; i++) {
 				JCheckBox button = new JCheckBox(jrbArray[i], false);
 				button.setActionCommand(action[i]);
 				buttonPanel.add(button);
